[+] Demonstration of runtime borrow checking.
Demo uses `RefCell<T>` to satisfy a trait method while maintaining
mutability at the cost of runtime borrow checking.

When a trait method mandates an immutable borrow of `self`, you can't
mutate `self` in a straightforward manner.

In the case you are a consumer of the trait, since you are not at
liberty to change the method signature, you must wrap the data you want
to mutate in a `RefCell<T>` and use it's `borrow_mut` method to grab a
mutable reference.